+++ kairo-serialization/src/test/kotlin/kairo/serialization/module/jvm/ValueClassDefaultObjectMapperTest.kt
@@ -1,0 +1,74 @@
+package kairo.serialization.module.jvm
+
+import com.fasterxml.jackson.databind.json.JsonMapper
+import com.fasterxml.jackson.module.kotlin.readValue
+import io.kotest.matchers.shouldBe
+import kairo.serialization.jsonMapper
+import kairo.serialization.serializationShouldFail
+import kairo.serialization.util.kairoWrite
+import kotlinx.coroutines.test.runTest
+import org.junit.jupiter.api.Test
+
+/**
+ * This test is intended to test behaviour strictly related to value class serialization/deserialization.
+ * Therefore, some test cases (such as unknown properties, pretty printing) are not included
+ * since they are not strictly related to value classes.
+ */
+internal class ValueClassDefaultObjectMapperTest {
+  @JvmInline
+  internal value class MyValueClass(val value: Int)
+
+  /**
+   * This test is specifically for non-nullable properties.
+   */
+  internal data class MyClass(
+    val value: MyValueClass,
+  )
+
+  private val mapper: JsonMapper = jsonMapper().build()
+
+  @Test
+  fun serialize(): Unit = runTest {
+    mapper.kairoWrite(MyClass(MyValueClass(42))).shouldBe("{\"value\":42}")
+  }
+
+  @Test
+  fun deserialize(): Unit = runTest {
+    mapper.readValue<MyClass>("{ \"value\": 42 }").shouldBe(MyClass(MyValueClass(42)))
+  }
+
+  @Test
+  fun `deserialize, null`(): Unit = runTest {
+    serializationShouldFail {
+      mapper.readValue<MyClass>("{ \"value\": null }")
+    }
+  }
+
+  @Test
+  fun `deserialize, missing`(): Unit = runTest {
+    serializationShouldFail {
+      mapper.readValue<MyClass>("{}")
+    }
+  }
+
+  @Test
+  fun `deserialize, wrong type, float`(): Unit = runTest {
+    serializationShouldFail {
+      mapper.readValue<MyClass>("{ \"value\": 1.23 }")
+    }
+  }
+
+  @Test
+  fun `deserialize, wrong type, string`(): Unit = runTest {
+    serializationShouldFail {
+      mapper.readValue<MyClass>("{ \"value\": \"42\" }")
+    }
+  }
+
+  @Test
+  fun `deserialize, wrong type, boolean`(): Unit = runTest {
+    serializationShouldFail {
+      mapper.readValue<MyClass>("{ \"value\": true }")
+    }
+  }
+}
